requirements.txt
flask
tensorflow==2.20.0
pillow
firebase-admin
requests
gunicorn
